Refactor Staking key and Stakepool dir structure
