+++ frontend/src/screens/Blockchain/Stakepool/BootstrapPool/index.js
@@ -1,29 +1,26 @@
 // @flow
-import idx from 'idx'
 import React, {useRef} from 'react'
 import {defineMessages} from 'react-intl'
+import idx from 'idx'
 import gql from 'graphql-tag'
 import {useQuery} from 'react-apollo-hooks'
-import useReactRouter from 'use-react-router'
 
-import {MetadataOverrides, seoMessages} from '@/pages/_meta'
-import {SummaryCard, SimpleLayout, LoadingInProgress} from '@/components/visual'
 import {LoadingError, EntityIdCard} from '@/components/common'
-
+import {SummaryCard, SimpleLayout, LoadingInProgress} from '@/components/visual'
 import {useScrollFromBottom} from '@/components/hooks/useScrollFromBottom'
-import {useI18n} from '@/i18n/helpers'
 import {useAnalytics} from '@/components/context/googleAnalytics'
-import Pool from '@/screens/Blockchain/StakingKey/Pool'
-
-const summaryLabels = defineMessages({
-  name: 'Name',
-  description: 'Description',
-})
+import {useI18n} from '@/i18n/helpers'
+import {MetadataOverrides, seoMessages} from '@/pages/_meta'
 
 const metadata = defineMessages({
   screenTitle: 'Slot Leader {poolHash} | Seiza',
   metaDescription: 'Cardano Slot Leader: {poolHash}. Name: {poolName}. Description: {description}.',
   keywords: 'Cardano Slot Leader {poolHash}, {commonKeywords}',
+})
+
+const summaryLabels = defineMessages({
+  name: 'Name',
+  description: 'Description',
 })
 
 const PoolSummaryCard = ({stakePoolData}) => {
@@ -69,12 +66,6 @@ const useStakePoolData = (poolHash) => {
   })
   return {loading, error, stakePoolData: data.stakePool}
 }
-
-// Note: We check directly for this hash (link from More screen)
-// to access mocked stakepool (non-bootstrap)
-// TODO: remove NON_BOOTSTRAP_POOL_HASH once we have real data
-export const NON_BOOTSTRAP_POOL_HASH =
-  'eccbc87e4b5ce2fe28308fd9f2a7baf3a87ff679a2f3e71d9181a67b7542122c'
 
 const SlotLeaderMetadata = ({poolHash, poolData}) => {
   const {translate: tr} = useI18n()
@@ -124,14 +115,4 @@ const BootstrapPool = ({poolHash}: {poolHash: string}) => {
   )
 }
 
-const StakePool = () => {
-  const {
-    match: {
-      params: {poolHash},
-    },
-  } = useReactRouter()
-  // TODO: remove NON_BOOTSTRAP_POOL_HASH once we have real data
-  return poolHash === NON_BOOTSTRAP_POOL_HASH ? <Pool /> : <BootstrapPool poolHash={poolHash} />
-}
-
-export default StakePool
+export default BootstrapPool
