+++ frontend/src/screens/Blockchain/Stakepool/NonBootstrapPool/Tabs/index.js
@@ -3,8 +3,8 @@ import {defineMessages} from 'react-intl'
 import {Grid} from '@material-ui/core'
 import {LiteTabs, LiteTab, LoadingInProgress} from '@/components/visual'
 import {useI18n} from '@/i18n/helpers'
-import History from '../../common/History'
-import Transactions from '../../common/Transactions'
+import History from '@/screens/Blockchain/common/History'
+import Transactions from '@/screens/Blockchain/common/Transactions'
 import {LoadingError, TabsPaginationLayout} from '@/components/common'
 import {useLoadStakepoolHistory} from '../dataLoaders'
