+++ frontend/src/screens/Blockchain/Stakepool/NonBootstrapPool/index.js
