+++ frontend/src/screens/Blockchain/StakingKey/Tabs/index.js
@@ -5,7 +5,7 @@ import {LiteTab, LiteTabs, LoadingInProgress} from '@/components/visual'
 import {LoadingError, TabsPaginationLayout} from '@/components/common'
 import {useI18n} from '@/i18n/helpers'
 import DelegatedPoolInfoTab from './DelegatedPoolInfoTab'
-import HistoryTab from '../../common/History'
+import HistoryTab from '@/screens/Blockchain/common/History'
 import Transactions from '../../common/Transactions'
 import {useLoadStakingKeyHistory} from '../dataLoaders'
