+++ frontend/src/screens/Blockchain/common/EmphasizedMessage.js
