+++ frontend/src/screens/Blockchain/common/mockedStakingData.js
