+++ frontend/src/screens/Blockchain/index.js
@@ -9,9 +9,8 @@ import Address from './Address'
 import Epoch from './Epoch'
 import Block from './Block'
 import PagedBlocks from './PagedBlocks'
-import StakePool from './StakePool'
-// TODO: refactor Staking key folders
-import StakingKey from './StakingKey/User'
+import StakePool from './Stakepool'
+import StakingKey from './StakingKey'
 import Slot from './Block/Slot'
 import NotFound from './NotFound'
 import SearchResults from './SearchResults'
